+++ libft/libft.h
@@ -16,7 +16,7 @@ void *ft_bzero(void *buf, size_t n);
 void *ft_memcpy(void *dst, const void *src, size_t n);
 void *ft_memmove(void *dest, const void *src, size_t n);
 size_t ft_strlcpy(char *dst, const char *src, size_t dstsize);
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size);
+unsigned int	ft_strlcat(char *dest, char *src, size_t size);
 int ft_toupper(int c);
 int ft_tolower(int c);
 char *ft_strchr(const char *s, int c);
